pyproject.toml
[project]
name = "ir-project"
version = "0.1.0"
description = "Implementation of the Build an IR system from Scratch project."
readme = "README.md"
requires-python = "~=3.10"
dependencies = [
    "python-terrier",
    "ir_datasets",
    "tqdm",
    "pandas",
    "more_itertools",
    "transformers>=4.36.0",
    "huggingface-hub>=0.36.0",
    "splade @ git+https://github.com/naver/splade.git",
    "pyt_splade @ git+https://github.com/cmacdonald/pyt_splade.git",
    "torch",
    "pandas-stubs~=2.3.3",
    "ipykernel>=7.1.0",
    "sentencepiece>=0.2.1",
    "pyterrier-t5>=0.2.1",
]

[tool.uv]
override-dependencies = [
    "transformers>=4.36.0",
    "tokenizers>=0.15.0",
]
